:construction: Refactor code
