+++ stopwatch/FeedbackModels.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+struct Question
+{
+    let text: String
+    let negativeButtonText: String
+    let positiveButtonText: String
+}
+
+struct Feedback
+{
+    let placeholderText: String
+    let negativeButtonText: String
+    let positiveButtonText: String
+}
